backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
akshare==1.18.30
openpyxl==3.1.2
xlsxwriter==3.1.9
openai==1.12.0
python-dotenv==1.0.0
tenacity==8.2.3
